chore: stop enforcing 100% coverage
There's apparently a bug on nyc (or istanbul) that prevents some lines to be marked as covered on windows, on OSX and linux everything seems to be working fine.

The lines marked as not covered on windows are confirmed to be running during the tests but for some reason they're not included on the test coverage.

This may be a bug on nyc and it was present before the node-tap update or v7 transports update. It's only getting attention now because the new node-tap versions enforce 100% coverage by default, which was making the windows CI tests to fail.